flight_agent/requirements.txt
flask
playwright
sib-api-v3-sdk
python-dotenv
gunicorn
